refactor demos for cleaner soc
